@@ -40,6 +40,7 @@
     "eslint-plugin-react": "^5.0.1",
     "express": "~4.14.0",
     "file-loader": "^0.9.0",
+    "html-webpack-plugin": "^2.26.0",
     "karma": "^0.13.22",
     "karma-mocha": "^0.2.2",
     "karma-phantomjs-launcher": "^1.0.0",
@@ -53,8 +54,8 @@
     "sinon": "^1.17.6",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.0",
-    "webpack-dev-server": "^1.14.1"
+    "webpack": "^2.2.0-rc.3",
+    "webpack-dev-server": "^2.2.0-rc.0"
   },
   "scripts": {
     "test": "karma start --single-run",
